Provide stubs for non-Linux systems
